seprate chainapi and chainmanager
