+++ tests/test_chain.py
@@ -63,8 +63,8 @@ def mine_next_block(parent, uncles=[], coinbase=None, transactions=[]):
         if utils.big_endian_to_int(o["result"]) <= 2**256 / b.difficulty:
             break
         else:
-            raise Exception("lol: "+str(b.difficulty))
-    assert b.header.check_pow()
+            raise Exception("lol: " + str(b.difficulty))
+    #assert b.header.check_pow()
     return b
 
 
@@ -350,7 +350,7 @@ def test_invalid_transaction():
 
 def test_prevhash():
     cm = new_chainmanager(db, mkquickgenesis({}))
-    L1 = mine_next_block(cm.head)
+    L1 = mine_next_block(cm.chain.head)
     L1.get_ancestor_list(2)
 
 
@@ -358,7 +358,7 @@ def test_genesis_chain():
     k, v, k2, v2 = accounts()
     db = new_db()
     blk = mkquickgenesis({v: {"balance": utils.denoms.ether * 1}})
-    chain = get_chainmanager(db=blk.db, genesis=blk)
+    chain = get_chainmanager(db=blk.db, genesis=blk).chain
 
     assert chain.has_block(blk.hash)
     assert blk.hash in chain
@@ -380,7 +380,7 @@ def test_simple_chain():
     k, v, k2, v2 = accounts()
     blk = mkquickgenesis({v: {"balance": utils.denoms.ether * 1}})
     store_block(blk)
-    chain = get_chainmanager(db=blk.db, genesis=blk)
+    chain = get_chainmanager(db=blk.db, genesis=blk).chain
     tx = get_transaction()
     blk2 = mine_next_block(blk, transactions=[tx])
     store_block(blk2)
@@ -435,16 +435,16 @@ def test_add_side_chain():
     cm = get_chainmanager(db=L0.db, genesis=L0)
     tx0 = get_transaction(nonce=0)
     L1 = mine_next_block(L0, transactions=[tx0])
-    cm.add_block(L1)
+    cm.chain.add_block(L1)
     tx1 = get_transaction(nonce=1)
     L2 = mine_next_block(L1, transactions=[tx1])
-    cm.add_block(L2)
+    cm.chain.add_block(L2)
 
     # receive serialized remote blocks, newest first
     transient_blocks = [rlp.decode(rlp.encode(R0), blocks.TransientBlock),
                         rlp.decode(rlp.encode(R1), blocks.TransientBlock)]
     cm.receive_chain(transient_blocks=transient_blocks)
-    assert L2.hash in cm
+    assert L2.hash in cm.chain
 
 
 def test_add_longer_side_chain():
@@ -468,17 +468,16 @@ def test_add_longer_side_chain():
     cm = get_chainmanager(db=L0.db, genesis=L0)
     tx0 = get_transaction(nonce=0)
     L1 = mine_next_block(L0, transactions=[tx0])
-    cm.add_block(L1)
+    cm.chain.add_block(L1)
     tx1 = get_transaction(nonce=1)
     L2 = mine_next_block(L1, transactions=[tx1])
-    cm.add_block(L2)
+    cm.chain.add_block(L2)
 
     # receive serialized remote blocks, newest first
     transient_blocks = [rlp.decode(rlp.encode(b), blocks.TransientBlock)
                         for b in remote_blocks]
     cm.receive_chain(transient_blocks=transient_blocks)
-    assert cm.head == remote_blocks[-1]
-
+    assert cm.chain.head == remote_blocks[-1]
 
 
 def test_reward_uncles():
@@ -495,21 +494,21 @@ def test_reward_uncles():
     uncle_coinbase = ('2' * 40).decode('hex')
     cm = get_chainmanager(db=blk0.db, genesis=blk0)
     blk1 = mine_next_block(blk0, coinbase=local_coinbase)
-    cm.add_block(blk1)
+    cm.chain.add_block(blk1)
     assert blk1.get_balance(local_coinbase) == 1 * blocks.BLOCK_REWARD
     uncle = mine_next_block(blk0, coinbase=uncle_coinbase)
-    cm.add_block(uncle)
-    assert uncle.hash in cm
-    assert cm.head.get_balance(local_coinbase) == 1 * blocks.BLOCK_REWARD
-    assert cm.head.get_balance(uncle_coinbase) == 0
+    cm.chain.add_block(uncle)
+    assert uncle.hash in cm.chain
+    assert cm.chain.head.get_balance(local_coinbase) == 1 * blocks.BLOCK_REWARD
+    assert cm.chain.head.get_balance(uncle_coinbase) == 0
     # next block should reward uncles
     blk2 = mine_next_block(blk1, uncles=[uncle.list_header()], coinbase=local_coinbase)
-    cm.add_block(blk2)
+    cm.cahin.add_block(blk2)
     assert blk2.get_parent().prevhash == uncle.prevhash
-    assert blk2 == cm.head
-    assert cm.head.get_balance(local_coinbase) == \
+    assert blk2 == cm.chain.head
+    assert cm.chain.head.get_balance(local_coinbase) == \
         2 * blocks.BLOCK_REWARD + blocks.NEPHEW_REWARD
-    assert cm.head.get_balance(uncle_coinbase) == blocks.BLOCK_REWARD * 7 / 8
+    assert cm.chain.head.get_balance(uncle_coinbase) == blocks.BLOCK_REWARD * 7 / 8
 
 
 # TODO ##########################################
